% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_language.R
\name{detect_language}
\alias{detect_language}
\title{This function will search through the Pandas DataFrame column of
textual data to detect the language of the corpus.}
\usage{
detect_language(df_col)
}
\arguments{
\item{df_col}{a list of values in data.frame column with textual data}
}
\value{
string the type of language present in the corpus
}
\description{
This function will search through the Pandas DataFrame column of
textual data to detect the language of the corpus.
}
\examples{
df = data.frame(text = c('He is a good guy.This is the worst coffee I had in my life.'))
detect_language(df$text_col)
}
