% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_of_speech.R
\name{get_part_of_speech}
\alias{get_part_of_speech}
\title{This function generates statistics about the proportions of following
parts of speech in the given column:
    - verbs
    - prepositions
    - adjectives
    - nouns
    - articles
    - etc.}
\usage{
get_part_of_speech(
  df_col,
  model_path = "./data/english-ewt-ud-2.4-190531.udpipe"
)
}
\arguments{
\item{df_col}{a list of values in data.frame column with textual data}

\item{model_path}{a string containing the path to the pretrained POS model.
     if provided, the function will look at it there, otherwise, the default place
is considered to be data/ folder. If the model doesn't exist, it will be downloaded.}
}
\value{
data.frame with one row containing columns verbs, prepositions,
 adjectives, nouns and articles.
}
\description{
This function generates statistics about the proportions of following
parts of speech in the given column:
    - verbs
    - prepositions
    - adjectives
    - nouns
    - articles
    - etc.
}
\examples{
ex <- data.frame(text_col =  c('Today is a beautiful Monday'))
get_part_of_speech(ex$text_col)
}
