% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarity.R
\name{polarity}
\alias{polarity}
\title{Polarity
  This method will check and compute the polarity of the text data.
  This method will return:
  - Number of Positive words
  - Number of Negative words}
\usage{
polarity(df_col)
}
\arguments{
\item{df_col}{This method will expect dataframe's column with textual data}
}
\value{
dataframe with one row containing columns for count of positive, negative and neutral words
}
\description{
Polarity
  This method will check and compute the polarity of the text data.
  This method will return:
  - Number of Positive words
  - Number of Negative words
}
\examples{
df = data.frame(text = c('He is a good guy.This is the worst coffee I had in my life.'))
polarity(df$text)
}
