% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_4.R
\name{summary_4}
\alias{summary_4}
\title{This function generates the following:
         - number of sentences,
         - number of stop words
         - list of high frequency
         - frequency of the words
      when a column of a dataframe or tibble is passed in.}
\usage{
summary_4(df_col)
}
\arguments{
\item{df_col}{(array): send in a column of the data.frame/tibble}
}
\value{
a data.frame with the following elements:
         nos    : number of sentences present in a column of the dataframe
         nosw   : number of stop-words present in a column of the dataframe
         list_hf: list of top 5 frequetly encountered words
         hf     : number of times the high frequency word was encountered.
}
\description{
This function generates the following:
         - number of sentences,
         - number of stop words
         - list of high frequency
         - frequency of the words
      when a column of a dataframe or tibble is passed in.
}
